Remove nomodeline, too contentious a default
This was copied from my vimrc without much thought to whether it would
be a friendly default, but it's probably too contentious to include.

modelines were historically susceptible to security vulnerabilities, but
this is no longer as much of a concern as no commands other than "set"
are supported, not all set commands are supported, and some set commands
run within a sandbox. Some distributions still disable modelines in the
system vimrc, and I personally recommend the securemodelines plugin, but
friendly.vim should not mess with modelines, they should work as normal.