Update wildmode to include full, more like zsh
The previous default was closely aligned with bash command completion,
requiring typing additional characters to continue partial completions.

Adding full allows tab completion of partial matches, more like zsh, and
most new users are likely to be using zsh, so this seems more friendly.

Possible TODO: check the user's default shell and set accordingly?
Problem with this is the best we can do is check the default shell, or
whatever they have set as &shell in vim, not the actual current shell.