Remap CTRL-U and CTRL-W to allow undo of changes
Remap to CTRL-U and CTRL-W to start a new change before deleting, which
allows the deletion to be undone. The Vim wiki has a good explanation:
https://vim.fandom.com/wiki/Recover_from_accidental_Ctrl-U

Both sensible.vim and Neovim defaults include both of these mappings,
Vim's own defaults.vim includes the CTRL-U mapping, which to be fair is
the more important one as it deletes more, so undo is more important.