Experimental mapping, CTRL-H to toggle hlsearch
Search highlighting can be confusing for new users, mostly because it's
not obvious how to turn it off after you're done with it (see :h nohls),
so friendly.vim automatically enables search highlighting while the
search cmdline is active and disables it when the search is complete.

As long as you're just searching for something this is fine, but if you
want to operate on the search matches using gn (see :h gn) then it's
very useful to be able to see them, so add a mapping to toggle hlsearch.

For new Vim users missing multiple cursors, gn is why you don't really
need multiple cursors in Vim, just operate on search matches instead :-)