Change synmaxcol from 200 to 500, fewer problems
Once synmaxcol is hit, syntax highlighting may be broken on subsequent
lines, so increase the limit to reduce the chances of breaking syntax
highlighting, but still avoid slow highlighting of super long lines.

Note: this was added to and then reverted from sensible.vim, because
it's not universally applicable (not everyone is writing code). See
https://github.com/tpope/vim-sensible/issues/142

Though this may cause problems for people writing documentation with
soft wrap enabled, I'd still prefer to make the default friendly for
writing code, and deal with other use cases separately (similar to
disabling soft wrap by default and then turning it on for markdown).