Replace fileignorecase with wildignorecase
Setting fileignorecase is probably not a good idea, Vim sets this
automatically on systems where case is normally ignored, e.g. MacOS,
but you may not want this on a case-sensitive filesystem where FOOBAR
and foobar are different files. Case-insensitive completion seems a
friendly default though, and that's exactly what wildignorecase does.