Add preview to completeopt, for ALEDocumentation
I generally have found using the preview window to show additional
information about completions really quite jarring, so I've always
disabled it, but I was tinkering with ALE to see if I could get away
with using it alone for LSP completions and documentation and found that
preview is necessary for ALEDocumentation to work properly (without it
ALEDocumentation ends up echoing most of the documentation rather than
writing it to the preview buffer, which is probably a bug, but an edge
case bug affecting people like me who remove preview from completeopt).

For simple keyword completions this has little affect anyway, there is
no additional info to show in the preview window, so this should only
affect people using omnicompletion, which I have not enabled in simple
tab completion because the default omnicompletions for some filetypes
are terrible (this is really something you want to opt in to).