Don't auto-complete ins mode completion matches
New users may not know how to cancel a completion, so auto-completing
the first or longest shared match can be really frustrating when you are
new to Vim, so require users to choose a match, which is more like other
editors or IDEs they might be familiar with (yep, I'm thinking vscode).

Experienced Vim users are likely to despise this, and I get it, it adds
keystrokes for every completion just to avoid confusion for novice users,
but this plugin is primarily targeted at novices, not experienced users.

This change also addresses a conflict between the previous completion
options which included menuone,longest and plugins like lexima that map
<CR> to accept completion when the menu is visible (when the longest
match was auto-completed it was not selected in the menu, so hitting
<CR> to accept completion went back to the original text, confusing).