Fix active window only cursorline for git commits
Don't set cursorline globally as this means the split for verbose git
commits gets created with a cursorline which then does not get hidden.

FIXME: make this feature aware of cursorline being set elsewhere, so it
only works as a default for new users, but is easily overridden without
resetting autocmds (or just remove it, it might be a bit much anyway).