Truncate long lines that cannot be displayed
Set display=truncate so that long lines that cannot be displayed within
the window are truncated and @@@ is used on the last line of the screen
to indicate the line was truncated to fit and more text is available.

This matters when wrap is enabled and long lines are continued on the
next line for display (friendly defaults to nowrap for most file types).