Fix = command indentation for text and commits
Re-indenting text using the = command falls back to using c-style
indentation when no file type specific indent has been specified.

This is weird with plain text and commit messages, for example it will
indent a line following a line with leading space and a trailing comma,
so add an indentexpr for text files and commits that is less weird.

For now I've limited this to text files and commits, but it may well
make a lot of sense for other non-code files like markdown and rdoc.

Note: "no file type specific indent" -> no indentexpr or equalprg set