Fix = command for mardown, asciidoc, rst, rdoc
Use the same basic indentexpr added for text and commits with other
documentation style file types so that c-style indentation is not
applied by default using the = command when editing these files.

This is a just default, a file type specific indentexpr or equalprg set
elsewhere, e.g. another vim plugin or in your vimrc, will override this.