Restore mappings for recording & replaying macros
Although q is easily hit by accident, and recording into a register can
be confusing for new users (certainly confused me when I started using
Vim), recording and replaying of macros is also a powerful feature and
I've concluded that disabling this by default is a really over the top.

This commit also remaps Q to repeat the last recorded register, which
is the same as the default in Neovim, and is easier to type than @@.

I am toying with the idea of mapping q to a no-op but allowing qq to
record the q register, thereby allowing recording of macros, but only
one at a time, which might be a reasonable compromise for new users.