List format aware indentexpr for non-code files
Restore the indentexpr for some non-code file types, to avoid c-style
indentation being applied when using the = command, but make it work
with list formatting applied via formatoptions and formatlistpat.

The indentation is only derived from formatlistpat if list formatting
has been enabled in formatoptions (i.e. formatoptions includes 'n'),
otherwise it falls back to previous line indent, which is still more
reasonable than c-style indent for non-code file types.

This works out of the box for markdown and gitcommit file types as the
ftplugins for these enable list formatting and set formatlistpat, but
not for text as the ftplugin uses comment formatting for bulleted lists.

To fix this, Friendly.vim now enables list formatting for text files,
and sets formatlistpat using the pattern from the gitcommit ftplugin.