Update README, add mouse support to feature list
It's not much of a feature, but it seems to be a common misconception
that the mouse is not supported by Vim, possibly because at least some
versions of Debian break defaults.vim by adding a one-line user vimrc,
for at least the root user, rather than adding to the system vimrc.

So, just make it clear that adding friendly.vim enables mouse support.