Hopefully smarter list formatting with friendly CR
- When deleting an empty list item on CR, and the previous line starts a
  list item, leave the cursor in position to continue the list item,
  same as when previous line itself is a continuation of a list item.
  Makes manual formatting easier with lines shorter than textwidth, hit
  CR once to start a new line with a new bullet, hit it again to remove
  the bullet and leave the cursor at the right position to continue.
- With the cursor on an empty line at the end of a list, get indent for
  the next line from the previous list header, not the previous line.
  This puts the cursor back at the previous indent after a list ends.
- Only delete empty list items on CR when previous line is non-blank.
  This allows for empty lists and other unexpected manual formatting.

These changes work well together which is why they are in one commit.
I won't be especially surprised if they come back to bite me and I end
up simplifying as there are edge cases that are too messy to deal with,
but it seems worth a go, I'll live with it a while and see how it works.

P.S. things get tricky with expandtab, friendly.vim sets noexpandtab by
default, but might be worth checking for expandtab and skipping if set.