Switch default color scheme from slate to habamax
I originally chose slate over habamax because it is included in Vim 8
and 9, but I didn't realise at the time that the version of slate in
Vim 8 is very basic, radically different to the version in Vim 9, and
not a whole lot better than the default color scheme (in my opinion).

The included color schemes were overhauled in Vim 9 / Neovim 0.8.0, so
check the Vim/Neovim version before overriding the default color scheme,
and use habamax over slate as it includes nice diff syntax highlighting.

It might be worth considering falling back to a slightly modified slate
for earlier Vim and Noevim versions. A couple of minor tweaks to comment
and diff highlighting would make the old version a pretty nice default.