Better solution to relative path completion
The previous solution originally came from my vimrc, and was cobbled
together years ago, before ++once was an option for Vim autocmds.

It also handled some edge cases which no longer seem necessary. The
WinLeave autocmd allowed changing the active window while the completion
menu was visible, but this doesn't seem necessary since an earlier move
from using InsertEnter and InsertLeave to CompleteDone. The VimLeavePre
autocmd was to handle quitting vim while the completion menu was visible
*and* automatically saving a session on exit, but this seems overkill as
the local working directory is only changed for file path completion,
and this is a hack for people saving sessions automatically on exit.

If you are automatically saving sessions on exit and you really want a
hack to make sure the working directory is reset you could try this:

    autocmd SessionLoadPost *
      \ windo if haslocaldir() | lcd - | cd | cd - | endif

The `cd` followed by `cd -` will clear the window local directory.