Always show results of search with grep and lgrep
The default behaviour when using grep or lgrep to search in files is to
populate the list of matches, jump to the first match if there is one,
and assume users are familiar with :copen, :cnext etc. to navigate.

This is confusing for new users as they can't see the list of matches,
so just open the quickfix or location list window to show the results.