Fix compatibility hack for commitia plugin
This appeared to work because committia removes the comments from the
commit message by the time friendly tries to extract the diff, but it
doesn't actually stop friendly from trying to extract the diff because
it tried to disable a VimEnter autocmd using a later VimEnter autocmd.

I've no idea how the sequence of autocmd events results in committia
deleting the diff before friendly looks for it, but that's what appears
to happen. It's possible something else is going on, may revisit later.