Set vert and eob fillchars to friendly defaults
Use same vert fillchar in Vim as Neovim, so when used to separate
vertically split windows it appears a solid rather than dashed line.

Set eob fillchar to space so empty lines below end of a buffer are just
empty. Line numbers are enabled, so it's obvious which lines are empty.