Remove preview from completeopt
This is included by default in both Vim and Neovim, and it can be very
useful, particularly for experienced users, but also very confusing for
new users - what is this window, why is it so tiny, how do I close it?

Experienced users will likely override this anyway, and will certainly
know how to, even relatively inexperienced users will likely override it
once they start tinkering with completion configurations and/or plugins.