Enable comment plugin on supported Vim versions
Included as an optional package in recent Vim versions, and a similar
feature is enabled by default in Neovim 0.10.0, so enable for Vim too.

This adds mappings for toggling comments in code, use gcc to toggle
comments on the current line, gc{motion} to comment a region, e.g. gcip.
Also works using visual mode to select and then gc to toggle comments.

See :h comment.txt for more details on Vim, :h commenting on Neovim